Fixed up the logic - use the openbabel target.
Using the CMake openbabel target, along with setting the third openbabel
variable, should allow all platforms to link. It also lets CMake track
the dependency, instead of adding it in manually if openbabel is
embedded. The target on Windows is openbabel-2, that should be changed
to be the same as all other platforms, even if the name must remain
different.